Added PostgreSQL support

+++ Server/Entities/ApplicationRole.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 namespace AspNetCoreSpa.Server.Entities
 {
     // Add profile data for application users by adding properties to the ApplicationUser class
-    public class ApplicationRole : IdentityRole<int>
+    public class ApplicationRole : IdentityRole<long>
     {
         [StringLength(250)]
         public string Description { get; set; }
