+++ Server/Entities/Content.cs
@@ -6,7 +6,7 @@ namespace AspNetCoreSpa.Server.Entities
     public class Content : IEntityBase
     {
         [Key]
-        public int Id { get; set; }
+        public long Id { get; set; }
         [Required]
         [StringLength(250)]
         public string Key { get; set; }
@@ -16,14 +16,14 @@ namespace AspNetCoreSpa.Server.Entities
 
     public class ContentText
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         [Required]
         [StringLength(2048)]
         public string Text { get; set; }
         public virtual Content Content { get; set; }
         public virtual Language Language { get; set; }
-        public int ContentId { get; set; }
-        public int LanguageId { get; set; }
+        public long ContentId { get; set; }
+        public long LanguageId { get; set; }
 
     }
